+++ src/components/profile/MyClass.tsx
@@ -48,15 +48,11 @@ export default function Class() {
                             ) {
                               console.log(e.currentTarget.name);
                               setIsBtnOpenList([
-                                isBtnOpenList[data.danceClassId] ==
-                                  !isBtnOpenList[data.danceClassId],
+                                !isBtnOpenList[data.danceClassId],
                                 ...isBtnOpenList,
                               ]);
                             } else {
-                              setIsBtnOpenList([
-                                isBtnOpenList[data.danceClassId] == false,
-                                ...isBtnOpenList,
-                              ]);
+                              setIsBtnOpenList([false, ...isBtnOpenList]);
                             }
                           }}
                           name={data.danceClassId.toString()}
@@ -101,7 +97,7 @@ export default function Class() {
                       </div>
                       <div className={styles.classDate}>
                         <Date className={styles.icon} />
-                        {data.method} {data.date}
+                        {data.method} · {data.date}
                       </div>
                     </div>
                   </div>
